+++ src/main/java/lotto/model/LottoService.java
@@ -1,0 +1,23 @@
+package lotto.model;
+
+public class LottoService {
+
+    private final WinningNumbers winningNumbers;
+    private final BonusNumber bonusNumber;
+
+    public LottoService(WinningNumbers winningNumbers, BonusNumber bonusNumber) {
+        validate(winningNumbers, bonusNumber);
+        this.winningNumbers = winningNumbers;
+        this.bonusNumber = bonusNumber;
+    }
+
+    private void validate(WinningNumbers winningNumbers, BonusNumber bonusNumber) {
+        boolean isDuplicate = winningNumbers.getWinningNumbers()
+                .stream()
+                .anyMatch(winningNumber -> winningNumber == bonusNumber.getBonusNumber());
+
+        if (isDuplicate) {
+            throw new IllegalArgumentException("[ERROR] 보너스 번호가 당첨 번호와 중복됩니다.");
+        }
+    }
+}
